backend/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-mail==1.4.2
h11==0.14.0
httpcore==1.0.7
httpie==3.2.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
multidict==6.1.0
natsort==8.4.0
passlib==1.7.4
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.2
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==76.0.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
